requirements.txt
Django==2.0
djangorestframework==3.8.2
mysqlclient==1.3.13
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
django-cors-headers==2.4.0
django-storages==1.6.6
boto3==1.8.4
Pillow
uwsgi
django-debug-toolbar==1.10.1
django-filter==2.0.0